Rename libgap to sage_interface
This is mainly done to avoid confusion over what "libgap"
is or should be.

The plan is to have an API to use GAP as a library in other
programs, and to achieve this we start out from libgap as
provided by Volker Braun and work towards an abstract interface
that will be useful to Sage as well as other projects.
