+++ src/gap.c
@@ -115,7 +115,7 @@
 #include <src/gaputils.h>
 
 #if defined(LIBGAP)
-#include <src/libgap.h>
+#include <src/sage_interface.h>
 #endif
 
 /****************************************************************************
