+++ src/gasman.h
@@ -1191,7 +1191,14 @@ extern void CallbackForAllBags(
 void *AllocateMemoryBlock(UInt size);
 #endif
 
-/* If not 0 this function will be called */
+/*
+ * If not 0 this function will be called in
+ * CollectBags to allow users of libgap to mark bags
+ */
 extern TExtraMarkFuncBags ExtraMarkFuncBags;
+
+static void SetExtraMarkFuncBags(TExtraMarkFuncBags func)
+{ ExtraMarkFuncBags = func; }
+
 
 #endif // GAP_GASMAN_H
