+++ src/sage_interface.c
@@ -17,11 +17,11 @@
  */
 
 
-#include <src/libgap.h>
+#include <src/sage_interface.h>
+#include <src/sage_interface_internal.h>
 #include <src/system.h>
 #include <src/gap.h>
 #include <src/gasman.h>
-#include <src/libgap_internal.h>
 #include <src/scanner.h>
 #include <assert.h>
 #include <stdio.h>
@@ -79,13 +79,7 @@ static libgap_gasman_callback_ptr gasman_callback = NULL;
 
 void libgap_set_gasman_callback(libgap_gasman_callback_ptr callback)
 {
-  gasman_callback = callback;
-}
-
-void libgap_call_gasman_callback()
-{
-  if (gasman_callback != NULL)
-    (*gasman_callback)();
+    SetExtraMarkFuncBags(callback);
 }
