+++ src/sysfiles.c
@@ -84,6 +84,9 @@ typedef void       sig_handler_t ( int );
 #include <process.h>
 #endif
 
+#if defined(LIBGAP)
+#include "sage_interface_internal.h"
+#endif
 
 /* utility to check return value of 'write'  */
 ssize_t writeandcheck(int fd, const char *buf, size_t count) {
